Split sage.py into multiple files
